requirements.txt
PyYAML==5.3.1
pystache==0.5.4
pysolr==3.8.1
GitPython==3.1.2
Flask==1.1.2
flask-restplus==0.13.0
Werkzeug==0.16.0
pika==1.1.0
apscheduler==3.6.3
